+++ src/@lekoarts/gatsby-theme-emma/@lekoarts/gatsby-theme-emma-core/components/project.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react';
+// import Project from "../../../components/project"
+import Project from '@lekoarts/gatsby-theme-emma/components/project';
+
+type Props = {
+	data: {
+		project: any;
+		[key: string]: any;
+	};
+	[key: string]: any;
+};
+
+export default function EmmaCoreProject({ ...props }: Props) {
+	props.data.project = props.data.mdxProject2;
+	return <Project {...props} />;
+}